UserGroupManagement/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
supabase==2.17.0
PyJWT==2.10.1
email_validator==2.2.0
pydantic==2.11.7
